adreno_tz: Adjust GPU target frequency calculation for HFR situations
Based on an idea by Sultan Alsawaf, however we do not have any way of
knowing which refresh rate the screen is currently running at. Most people
are using 90HZ, so multiply the value by 1.5.

original commit message below:

It seems that the black-box GPU frequency selection algorithm is unaware
of high refresh rates, and assumes a refresh rate of 60 Hz for its
target frequencies. This results in frequencies that are too low being
selected when a higher refresh rate, such as 90 Hz, is used.

To remedy this, multiply the GPU's busy time by 1.5 when 90 Hz is used,
since the graphical load of 90 Hz should be about 1.5x greater than the
load of 60 Hz. This makes msm-adreno-tz select the right frequencies
needed to avoid stuttering.

Signed-off-by: Yaroslav Furman <yaro330@gmail.com>